@@ -3,14 +3,14 @@
   "version": "1.3.1",
   "description": "JWT utils for Fastify",
   "main": "jwt.js",
-  "types": "index.d.ts",
+  "types": "jwt.d.ts",
   "engines": {
-    "node": ">6"
+    "node": ">10"
   },
   "scripts": {
     "lint": "standard",
     "lint:fix": "standard --fix",
-    "test": "standard && tap test.js && tsc --project ./tsconfig.json"
+    "test": "standard && tap test.js && tsd"
   },
   "repository": {
     "type": "git",
@@ -37,15 +37,11 @@
     "steed": "^1.1.3"
   },
   "devDependencies": {
-    "fastify": "^2.0.0",
+    "fastify": "^3.0.0-alpha.1",
     "fastify-cookie": "^3.5.0",
     "standard": "^14.0.2",
     "tap": "^12.6.5",
+    "tsd": "^0.11.0",
     "typescript": "^3.2.2"
-  },
-  "greenkeeper": {
-    "ignore": [
-      "tap"
-    ]
   }
 }
